+++ example/jest.config.js
@@ -1,3 +1,4 @@
 module.exports = {
-  preset: 'react-native',
+  preset: '@react-native/jest-preset',
+  watchman: false,
 };